Implemented parseLatLon.
Implemented the parseLatLon function in the LatLon class. It accepts a
coordinate String value as input and converts it to a LatLon instance.
The functionality was harvested from the GoToCoordinatePanel class which
implemented similar functionality in its compuateLatLonFromString
function. Refactored the GoToCoordinatePanel to make use of LatLon's
parseLatLon function instead.