Added arrowOffset parameter.
Introduced a new parameter, called arrowOffset. It is used as a ratio to
specify the offset of the arrow from the first to the last position
where a value of 0.0 indicates that the arrow is on the starting
position and a value of 1.0 indicates that the arrow is on the last
position. The default is 0.5 which specifies that the arrow is halfway
between the two positions.