Small changes to MeasureTool.
Made some small changes to the MeasureTool. In particular if the
shape-type was set to PATH and we seed the MeasureTool with positions
then the shape-type should stay set to PATH. The old logic caused the
MeasureTool to auto-detect the number of positions and then set the
shape-type appropriately which is not what we want in most cases.