Shapefile borders and coastlines layers.
Added two new classes called CoastlinesLayer and BordersLayer. These
classes extend from ShapefileLayer and loads their shapes from
Coastlines and Borders shapefiles respectively. The purpose of these two
layers is to provide an offline alternative to the
CountryBoundariesLayer. The only downside is that these layers might be
a bit more resource intensive to render and they are a bit lower
resolution when compared to the image-based CountryBoundariesLayer.
Also, it inflates the ww-core jar a bit.