Refactored BasicElevationModel.
Removed the computeLevelForResolution() and getTilesInSector() methods
from the BasicElevationModelBulkDownloader class and moved the
functionality over to the BasicElevationModel class itself. This is to
correspond to the BasicTiledImageLayerBulkDownloader class that calls
the TiledImageLayer's computeLevelForResolution() and getTilesInSector()
methods. The advantage of this refactor is that given the
elevation-layer we can now get all the levels and tiles, whereas in the
past we could only do this for image-layers.