Changed versioning to be SNAPSHOT.
Changed the versioning to be 2.0.0-SNAPSHOT. This allows snapshot
releases to be deployed to a nexus repository with an automatic
timestamp attached to the version. This basically mirrors the
functionality of the NASA daily WorldWind builds. Except, now you can
perform the daily build yourself and deploy to your own internal
repository.