Zero-degree fix for Angle class.
When an angle is created via the fromDMS() method and a 0-degree
parameter is specified, the internal degree-calculation is incorrect
in that the sign then becomes 0. This results in the minutes and
seconds being multiplied with 0 as well, with the end result being a
0 decimal degrees angle. This is now fixed in that the sign falls back
to a positive 1.0 multiplier when the degrees parameter is 0. A
unit-test was added to check that this boundary condition is correctly
handled.