Added additional Position constructor methods.
Added additional static constructor methods to the Position class. In
particular an additional fromRadians() method was added that takes only
the latitude and longitude as parameters (this mirrors the fromDegrees()
method that also only takes the latitude and longitude as parameters).
Then, added two methods namely fromLatLon() to construct a Position
class from a given LatLon object. The first method (with the elevation
as parameter) is a convenience method that mirrors the second
constructor. The second method (without the elevation as parameter) is
more useful as it constructs a Position object from a given LatLon
instance and assigns a default elevation of 0 for it.