Fixed Path selection.
Modified the Path class' getExtent() method to return a bounding-box
around its computed sector (via the getSector() method). This fixes the
selection problem in that the correct extent used to determine if the
path should be considered for picking or not. Removed the redundant
computeExtent() method and replaced calls to it with calls to the new
getExtent() method.