ByteBuffer adds support for extracting numbers and arrays from an underlying ArrayBuffer, such as what might be returned by XMHHttpRequest (xhr).
Arrays are extracted so as to reused the underlying ArrayBuffer and not allocate additional memory other than for the wrapper of the array.

The exception to this is for Float64Array, which must be aligned on an 8-byte boundary. If the underlying data for Float64Array is not 8-byte aligned, an intermediate array is allocated, which will be 8-byte aligned.

git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@2954 822214bb-dcf7-4c8d-a767-17ae0fb20d88