Fix 2 picking bugs with SurfaceShapes.
Previously, it was possible for a SurfaceShape to have several pick colors associated with it, resulting in the shape appearing in the pick list multiple times. This change assures that there is only ever a single pick color for any SurfaceShape.

The second issue addressed is forcing the creation of a texture for a SurfaceShapeTile when in picking mode. Since pick colors change dynamically, the texture can't be cached.

git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@3006 822214bb-dcf7-4c8d-a767-17ae0fb20d88